hook server up to `semgrep mcp` backend
